feat: auth callback + middleware
